Added Basic AI
Added basic slime AI that follows a preset path
Fixed collision detection.

+++ HAPI_Start/DamageComponent.hpp
@@ -6,7 +6,7 @@ struct DamageComponent : public Component<DamageComponent> {
 	DamageComponent() {}
 	std::shared_ptr<BaseComponent> Clone() const override;
 
-	int Damage;
+	float Damage;
 	bool DestroyOnHit{ false };
 	std::string EntityToSpawnOnHit;
 };
